Explicitly declare top-level Wordpress class
Thanks to an annoying bug in Puppet, if you're declaring a top-level
class inside a namespace that shares the same name (i.e. we're in
profiles::wordpress and trying to declare the wordpress class from
the wordpress module), then you have to fully namespace the declaration
(like ::wordpress) or Puppet will think it's a relative declaration
to the current class you're in (i.e. profiles::wordpress).

This is annoying, but necessary.