0.0.2 - include a wordpress_localhost class
Previously, the wordpress class assumed name-based lookup and setting
up a host entry for the node. Because I'm preparing for a Vagrant demo
which will all be self-contained, I created a class called wordpress_localhost
that will define everything on localhost and also setup a firewall rule to
allow all traffic on port 80.  I've also updated the Modulefile to add
the dependency on the puppetlabs/firewall module.