Stupid namespacing + firewall
So, I HATE that Puppet namespacing bug - unless I say ::jenkins,
it's going to relatively declare a jenkins class...in this case, our
class is called 'profiles::jenkins' and an 'include jenkins' is going
to look relative to the profiles class and try to declare the current
class we're in...hence, profiles::jenkins.  When I say 'include ::jenkins'
it will look for a jenkins class in top scope.  Scumbag Puppet.

Also, let's add a Firewall rule for port 8080 just to have fun.